Fix building with 5.5 again
Specifically, guarding the APIs that don't exist in 5.5, and fixing the things
that keep the codebase from building with 5.5 outside of those APIs; namely,
5.6 removes a lot of const qualifiers, which is good, but no help to those
building with 5.5. The affected text APIs almost all work on static buffers
anyways, so I just pass the returned buffers directly into the String
constructor (which doesn't care about the const-ness of the `char*`) and call
it a day. As for the build issues that can't be fixed that way, I changed the
variable holding the return of TextSplit to be `const char**` (since we don't
modify the value anyway, not that we should since it's a static buffer), and I
added a little stub to convert the `char**` argument for TextJoin to be
`const char**`, which can be deleted easily once the next version of raylib
comes out.